+++ core/src/main/java/com/tilldawn/Model/monster/WingedMonster.java
@@ -95,7 +95,7 @@ private float lastDashTime = 0f;
 
 
 
-    public void takeDamage(int dmg) {
+    public void takeDamage(float dmg) {
         if (!dying) {
             health -= dmg;
             if (health <= 0) {
